fix: Refactored the execution into multiple files
